everything is done except for update version

+++ Bizsolv Manpower Costing/src/model/Computation.java
@@ -3,9 +3,9 @@ package model;
 public abstract class Computation {
 
     private int workingDays;
-    private double basicSalary, totalGovernmentalCost, totalLaborCost, bizsolvAdminCost, contractCost;
+    private float basicSalary, totalGovernmentalCost, totalLaborCost, bizsolvAdminCost, contractCost;
 
-    public Computation(double basicSalary, int workingDays) {
+    public Computation(float basicSalary, int workingDays) {
         this.basicSalary = basicSalary;
         this.workingDays = workingDays;
     }
@@ -18,38 +18,38 @@ public abstract class Computation {
         this.workingDays = workingDays;
     }
 
-    public double getBasicSalary() {
+    public float getBasicSalary() {
         return basicSalary;
     }
 
-    public void setBasicSalary(double basicSalary) {
+    public void setBasicSalary(float basicSalary) {
         this.basicSalary = basicSalary;
     }
 
-    public double getMonthBonus() {
+    public float getMonthBonus() {
         return basicSalary/12;
     }
 
-    public void setGovernmentalCost(double value) { this.totalGovernmentalCost = value; }
+    public void setGovernmentalCost(float value) { this.totalGovernmentalCost = value; }
 
-    public double getTotalGovernmentalCost() { return totalGovernmentalCost; }
+    public float getTotalGovernmentalCost() { return totalGovernmentalCost; }
 
-    public void setTotalLaborCost(double value) { this.totalLaborCost = value; }
+    public void setTotalLaborCost(float value) { this.totalLaborCost = value; }
 
-    public double getTotalLaborCost() { return this.totalLaborCost; }
+    public float getTotalLaborCost() { return this.totalLaborCost; }
 
-    public void setBizsolvAdminCost(double value) { this.bizsolvAdminCost = value; }
+    public void setBizsolvAdminCost(float value) { this.bizsolvAdminCost = value; }
 
-    public double getBizsolvAdminCost() { return this.bizsolvAdminCost; }
+    public float getBizsolvAdminCost() { return this.bizsolvAdminCost; }
 
-    public void setContractCost(double value) { this.contractCost = value; }
+    public void setContractCost(float value) { this.contractCost = value; }
 
-    public double getContractCost() { return this.contractCost; }
+    public float getContractCost() { return this.contractCost; }
 
-    abstract public double getNumOfDayIncentive();
-    abstract public double getEquivalentMonthlyCost();
-    abstract public double getEffectiveMonthlyRate();
-    abstract public double getAllowance();
-    abstract public double getSubTotal();
-    abstract public double getAdminCost();
+    abstract public float getNumOfDayIncentive();
+    abstract public float getEquivalentMonthlyCost();
+    abstract public float getEffectiveMonthlyRate();
+    abstract public float getAllowance();
+    abstract public float getSubTotal();
+    abstract public float getAdminCost();
 }
